リファクタ_UserControllerに関するファイルをメンテナンス
