+++ web/resources/views/layouts/app.blade.php
@@ -59,7 +59,7 @@
                                 </a>
 
                                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdown">
-                                    <a class="dropdown-item" href="{{ route('users.show', ['user' => Auth::user()->id]) }}">
+                                    <a class="dropdown-item" href="{{ route('users.show') }}">
                                         {{ __('view.MyPage') }}
                                     </a>
                                     <div class="dropdown-divider"></div>
